restructured folder organization
