+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sagansat-website",
-  "version": "0.1.0",
+  "version": "1.0.1",
   "license": "MIT",
   "scripts": {
     "dev": "next dev",
