+++ src/components/Help/Header/styles.ts
@@ -2,7 +2,7 @@ import { lighten } from 'polished'
 import styled from 'styled-components'
 
 export const Container = styled.header`
-  background: linear-gradient(to bottom, rgba(0, 0, 0, 0.3) 0%, rgba(0, 0, 0, 0.7) 75%, #000000 100%), url('/bg-help.jpg');
+  background: linear-gradient(to bottom, rgba(0, 0, 0, 0.3) 0%, rgba(0, 0, 0, 0.7) 75%, #000000 100%), url('/images/bg-help.jpg');
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
