+++ src/components/Layout/Footer/ContactUs/styles.ts
@@ -11,7 +11,7 @@ export const Container = styled.section`
   background:
     linear-gradient(to top, rgba(0, 0, 0, 0.1) 0%, rgba(0, 0, 0, 0.5) 75%, #000000 100%),
     linear-gradient(to bottom, rgba(0, 0, 0, 0.1) 0%, rgba(0, 0, 0, 0.5) 75%, #000000 100%),
-    url('/bg-contact-us.jpg');
+    url('/images/bg-contact-us.jpg');
   background-position: center;
   background-repeat: no-repeat;
   background-attachment: scroll;
