--- src/components/Projects/styles.ts
@@ -1,15 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.section`
-  background: var(--background);
-`
-
-export const Content = styled.div`
-  max-width: 1120px;
-  margin: 0 auto;
-  padding: 5.5rem 2rem;
-
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-`
