+++ src/pages/research/styles.ts
@@ -1,6 +1,8 @@
 import styled from 'styled-components'
 
-export const Container = styled.section`
+const Container = styled.section`
   display: flex;
   flex-direction: column;
 `
+
+export default Container
